Break NetworkManager out into separate file
