+++ DarkSky-Practica-A2/Services/WeatherService.swift
@@ -56,24 +56,3 @@ class WeatherSerivce {
     }
 }
 
-class NetworkManager {
-    
-    class func getDataFor(url: URL, completion: @escaping (Data) -> ()) {
-        
-        let config = URLSessionConfiguration.default
-        let session = URLSession(configuration: config)
-        let request = URLRequest(url: url)
-        
-        session.dataTask(with: request) { (data, response, error) in
-            guard let responseData = data else {
-                print("could not gather data")
-                return
-            }
-            print(response.debugDescription)
-            completion(responseData)
-            
-        }.resume()
-        
-    }
-    
-}
